added base project readme  files
base project files mostly contain the readme files that will be helpful in showing the requisites of the project and also to give a sense of direction on are we doing and hat needs to be done. A quick reado on the readme files would be appreciated and also helpful so that we can be on the same page.